Add validations for commands

+++ src/MoneyUnderControl.Domain/Commands/ExpenseReportCommandHandler.cs
@@ -1,10 +1,10 @@
-﻿using System;
-using FluentValidation.Results;
+﻿using FluentValidation.Results;
 using MediatR;
-using System.Threading;
-using System.Threading.Tasks;
 using MoneyUnderControl.Domain.Entities;
 using MoneyUnderControl.Domain.Interfaces;
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace MoneyUnderControl.Domain.Commands
 {
@@ -22,15 +22,19 @@ namespace MoneyUnderControl.Domain.Commands
 
         public async Task<ValidationResult> Handle(RegisterNewExpenseReportCommand request, CancellationToken cancellationToken)
         {
+            if (!request.IsValid()) return request.ValidationResult;
+
             var expense = new ExpenseReport(Guid.NewGuid(), request.Item, request.Price, request.ExpenseDate, DateTime.Now, request.Category, request.Status, request.Description);
 
-            _repository.Add(expense);
+            _repository.Add(expense); 
 
             return new ValidationResult();
         }
 
         public async Task<ValidationResult> Handle(UpdateExpenseReportCommand request, CancellationToken cancellationToken)
         {
+            if (!request.IsValid()) return request.ValidationResult;
+
             var expense = new ExpenseReport(request.Id, request.Item, request.Price, request.ExpenseDate, request.ReleaseDate, request.Category, request.Status, request.Description);
 
             _repository.Update(expense);
@@ -40,6 +44,8 @@ namespace MoneyUnderControl.Domain.Commands
 
         public async Task<ValidationResult> Handle(RemoveExpenseReportCommand request, CancellationToken cancellationToken)
         {
+            if (!request.IsValid()) return request.ValidationResult;
+
             var expense = new ExpenseReport(request.Id, request.Item, request.Price, request.ExpenseDate, request.ReleaseDate, request.Category, request.Status, request.Description);
 
             _repository.Remove(expense);
